Inline Pact Library pipeline
